requirements.txt
Django==4.1.5
oauth2==1.9.0.post1
psycopg2==2.9.5
djangorestframework==3.14.0
pytz==2022.7.1
django-cors-headers==3.13.0
aiohttp==3.8.3
aiosignal==1.3.1 
async-timeout==4.0.2 
attrs==22.2.0
base58==2.1.1
bitarray==2.6.2
charset-normalizer==2.1.1
cytoolz==0.12.1
eth-abi==4.0.0b2
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.2.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
multiaddr==0.0.9
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.21.12
pycryptodome==3.17
pyrsistent==0.19.3
rlp==3.0.0
six==1.16.0
toolz==0.12.0
web3==6.0.0b9
websockets==10.4
yarl==1.8.2
lru-dict==1.1.8
regex==2022.10.31
PyJWT==2.6.0